Website/backend/requirements.jetson.txt
# =========================
# Jetson / Local full stack (heavy ML deps)
# =========================

# Core API
fastapi==0.123.5
uvicorn==0.38.0
python-multipart==0.0.9
python-dotenv==1.2.1
requests==2.32.5
pydantic==2.12.5
typing-extensions>=4.8.0

# Auth / DB
SQLAlchemy==2.0.44
greenlet==3.2.4
bcrypt==5.0.0
passlib==1.7.4

# Document Processing
pypdf==4.2.0

# IMPORTANT — keep numpy < 2
numpy==1.26.4

# Camera / CV
opencv-python-headless==4.10.0.84

# LangChain / RAG
langchain==0.2.14
langchain-core==0.2.36
langchain-community==0.2.12
langchain-text-splitters==0.2.2
langchain-ollama==0.1.3
lightrag==0.1.0b6
faiss-cpu==1.8.0.post1
rank-bm25==0.2.2
websockets==15.0.1

# Extras
flask==3.0.3
faster-whisper==1.0.3
pyttsx3==2.90
ultralytics==8.3.0